+++ lib/haml_lint/ruby_extractor.rb
@@ -113,8 +113,8 @@ module HamlLint
         end
       else
         add_line('puts', node)
-        HamlLint::Utils.extract_interpolated_values(node.text) do |interpolated_code|
-          add_line(interpolated_code, node)
+        HamlLint::Utils.extract_interpolated_values(node.text) do |interpolated_code, line|
+          add_line(interpolated_code, node.line + line)
         end
       end
     end