+++ lib/haml_lint/utils.rb
@@ -20,17 +20,44 @@ module HamlLint
       end
     end
 
-    # Yields interpolated values within a block of filter text.
-    def extract_interpolated_values(filter_text)
-      Haml::Util.handle_interpolation(filter_text.dump) do |scan|
+    # Yields interpolated values within a block of text.
+    #
+    # @param text [String]
+    # @yield Passes interpolated code and line number that code appears on in
+    #   the text.
+    # @yieldparam interpolated_code [String] code that was interpolated
+    # @yieldparam line [Integer] line number code appears on in text
+    def extract_interpolated_values(text) # rubocop:disable Metrics/AbcSize
+      dumped_text = text.dump
+      newline_positions = extract_substring_positions(dumped_text, '\\\n')
+
+      Haml::Util.handle_interpolation(dumped_text) do |scan|
+        line = (newline_positions.find_index { |marker| scan.pos <= marker } ||
+                newline_positions.size) + 1
+
         escape_count = (scan[2].size - 1) / 2
-        return unless escape_count.even? # rubocop:disable Lint/NonLocalExitFromIterator
+        break unless escape_count.even?
 
         dumped_interpolated_str = Haml::Util.balance(scan, '{', '}', 1)[0][0...-1]
 
         # Hacky way to turn a dumped string back into a regular string
-        yield eval('"' + dumped_interpolated_str + '"') # rubocop:disable Eval
+        yield [eval('"' + dumped_interpolated_str + '"'), line] # rubocop:disable Eval
       end
+    end
+
+    # Returns indexes of all occurrences of a substring within a string.
+    #
+    # Note, this will not return overlaping substrings, so searching for "aa"
+    # in "aaa" will only find one substring, not two.
+    #
+    # @param text [String] the text to search
+    # @param substr [String] the substring to search for
+    # @return [Array<Integer>] list of indexes where the substring occurs
+    def extract_substring_positions(text, substr)
+      positions = []
+      scanner = StringScanner.new(text)
+      positions << scanner.pos while scanner.scan(/(.*?)#{substr}/)
+      positions
     end
 
     # Converts a string containing underscores/hyphens/spaces into CamelCase.